scmutationexec: break down scmutationexec.go
This refactoring commit breaks down the MutationVisitor implementation
defined in scmutationexec.go into multiple source files with distinct
themes: index operations in index.go, column ops in column.go, etc...

Release note: None
